Algumas correções

+++ src/main/java/sistema/dao/ClienteDAO.java
@@ -3,6 +3,6 @@ package sistema.dao;
 import sistema.dao.generic.DAOImpl;
 import sistema.modelos.Cliente;
 
-public class ClienteDAO extends DAOImpl<Cliente, Integer>{
+public class ClienteDAO extends DAOImpl<Cliente, Long>{
 
 }
